requirements.txt
openai
click
python-dotenv
pytest
pylint
black
